ecommerce-backend/pyproject.toml
[project]
name = "ecommerce-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.7"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.30.0",
    "fastapi-users[oauth,sqlalchemy]>=15.0.1",
    "fastapi[standard]>=0.121.2",
    "pydantic-settings>=2.12.0",
]

[tool.black]
line-length = 80
include = '\.pyi?$'
extend-exclude = '''
/(
  # The following are specific to Black, you probably dont want those.
  | blib2to3
  | tests/data
  | profiling
  | migrations
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
skip = [".gitignore", "migrations"]

[tool.pylint.MESSAGE_CONTROL]
disable = [
    "cyclic-import",
    "duplicate-code",
    "invalid-name",
    "line-too-long",
    "missing-class-docstring",
    "missing-module-docstring",
    "missing-function-docstring",
    "too-few-public-methods",
]
enable = "useless-suppression"
extension-pkg-allow-list = ["pydantic", "asyncpg.pgproto.pgproto", "orjson"]

[tool.pylint.VARIABLES]
ignored-argument-names = "args|kwargs|_|__"

[tool.pylint.BASIC]
good-names = "_,__,i,e,k,v,fn,get,post,put,patch,delete,route,asgi,websocket,Dependency,Body,Parameter"
# Regular expression which should only match function or class names that do
# not require a docstring.
no-docstring-rgx = "(__.*__|main|test.*|.*test|.*Test|^_.*)$"

[tool.pylint.REPORTS]
reports = "no"

[tool.pylint.FORMAT]
max-line-length = "80"

[tool.pylint.LOGGING]
# Logging modules to check that the string format arguments are in logging
# function parameter format
logging-modules = ["logging"]

[dependency-groups]
dev = [
    "aiosqlite>=0.21.0",
    "black>=25.11.0",
    "faker>=38.2.0",
    "httpx>=0.28.1",
    "isort>=7.0.0",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
]


[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.pyrefly]
ignore = ["*/__pycache__", "*/migrations"]
